Fix #217, use size_t in PSP stubs
Recent API updates require use of the size_t type
instead of uint32.

+++ unit-test-coverage/ut-stubs/src/libc-stdlib-stubs.c
@@ -85,7 +85,7 @@ void *PCS_malloc(size_t sz)
     cpuaddr PoolEnd;
     cpuaddr NextBlock;
     size_t NextSize;
-    uint32 PoolSize;
+    size_t PoolSize;
     uint32 CallCnt;
     struct MPOOL_REC *Rec;
 
@@ -156,7 +156,7 @@ void PCS_free(void *ptr)
     int32 Status;
     cpuaddr BlockAddr;
     void *PoolPtr;
-    uint32 PoolSize;
+    size_t PoolSize;
     struct MPOOL_REC *Rec;
 
     /*
